Fix build on FreeBSD (#585)
FreeBSD make and GNU make share a common base but provide extensions
not supported by the other.  This leads to build failures and can be
fixed by using the right version of make.  On FreeBSD, GNU make is
installed as `gmake` so detect and use this.  While here, also check
the environment for allowing overriding the binary to make in case
other platform can benefit from it (e.g. OpenBSD, NetBSD and probably
others).

While here, also catch the right number of CPU for parallel build.